ça bouge !!!!!!!!! et y a des plans !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!

+++ Solution/Math_Physique_JV_ProjetSession/Sources/Physic/ForceGenerator/ParticleForceGenerator.h
@@ -7,6 +7,6 @@ class ParticleForceGenerator
 
 {
 public:
-	virtual void UpdateForce(Particule* particle, float duration) = 0;
+	virtual void UpdateForce(Particule* particle, double duration) = 0;
 };
